move out relationship new contracts to v1
