+++ app/api/relationships/query/express/GetRelationshipsAnchorsController.ts
@@ -1,5 +1,5 @@
 import { AbstractController } from '#api/common.v2/infrastructure/AbstractController.js';
-import { RelationshipsQueryServiceFactory } from '#api/core/infrastructure/factories/RelationshipsQueryServiceFactory.js';
+import { RelationshipsQueryServiceFactory } from '#api/relationships/query/factory/RelationshipsQueryServiceFactory.js';
 import { AnchorsQuerySchema } from './RelationshipQuerySchemas.js';
 import { sendNotFoundRows } from './sendNotFoundRows.js';
 
@@ -7,7 +7,7 @@ class GetRelationshipsAnchorsController extends AbstractController {
   protected async handle(): Promise<void> {
     try {
       const query = AnchorsQuerySchema.parse(this.request.query);
-      const rows = await RelationshipsQueryServiceFactory.default(this.user).getAnchors({
+      const rows = await RelationshipsQueryServiceFactory.default().getAnchors({
         ...query,
         language: this.language,
       });
